上传、展示所有文件完成

+++ src/main/java/com/nk/entity/Myfile.java
@@ -12,7 +12,7 @@ public class Myfile {
 
     private String cdate;
 
-    private Double fsize;
+    private Long fsize;
 
     private String ftype;
